@@ -18,6 +18,7 @@
     "karma-lib-prod": "bash -c \"(export NODE_ENV=production; npm run karma-lib -loglevel silent)\"",
     "eslint-internal": "eslint --cache ./src/ ./docs/src/ ./tests/src/ --ext .js --ext .jsx",
     "eslint": "npm run eslint-internal -loglevel silent || true",
+    "eslint-fix": "eslint --cache ./src/ ./docs/src/ ./tests/src/ --ext .js --ext .jsx --fix || true",
     "tdd": "bash -c \"(export KARMA_TDD=true; npm run karma)\"",
     "tdd-prod": "bash -c \"(export KARMA_TDD=true; npm run karma-prod)\"",
     "cover": "bash -c \"(export KARMA_COVERAGE=true; npm run karma-src)\""
@@ -52,11 +53,11 @@
     "chai": "3.5.0",
     "del": "2.2.2",
     "dirty-chai": "1.2.2",
-    "eslint": "3.10.2",
-    "eslint-config-airbnb": "10.0.1",
-    "eslint-plugin-import": "2.0.1",
-    "eslint-plugin-jsx-a11y": "^2.1.0",
-    "eslint-plugin-react": "6.4.1",
+    "eslint": "3.11.0",
+    "eslint-config-airbnb": "13.0.0",
+    "eslint-plugin-import": "2.2.0",
+    "eslint-plugin-jsx-a11y": "2.2.3",
+    "eslint-plugin-react": "6.7.1",
     "estraverse-fb": "1.3.1",
     "ghooks": "1.3.2",
     "gulp": "3.9.1",
@@ -73,7 +74,7 @@
     "karma-spec-reporter": "0.0.26",
     "karma-travis-fold-reporter": "1.0.1",
     "karma-webpack": "1.8.0",
-    "mocha": "3.1.2",
+    "mocha": "3.2.0",
     "mocha-loader": "1.0.0",
     "pngjs2": "2.0.0",
     "react": "15.4.0",
