@@ -163,9 +163,9 @@ class React3CompositeComponentWrapper extends ReactCompositeComponentMixinImpl {
    * @internal
    */
   mountComponent(transaction,
-                 hostParent,
-                 hostContainerInfo,
-                 context) {
+    hostParent,
+    hostContainerInfo,
+    context) {
     this._context = context;
     this._mountOrder = this._react3RendererInstance.nextMountID++;
     this._hostParent = hostParent;
@@ -336,9 +336,9 @@ class React3CompositeComponentWrapper extends ReactCompositeComponentMixinImpl {
   }
 
   _constructComponent(doConstruct,
-                      publicProps,
-                      publicContext,
-                      updateQueue) {
+    publicProps,
+    publicContext,
+    updateQueue) {
     if (process.env.NODE_ENV !== 'production') {
       ReactCurrentOwner.current = this;
       try {
@@ -362,9 +362,9 @@ class React3CompositeComponentWrapper extends ReactCompositeComponentMixinImpl {
   }
 
   _constructComponentWithoutOwner(doConstruct,
-                                  publicProps,
-                                  publicContext,
-                                  updateQueue) {
+    publicProps,
+    publicContext,
+    updateQueue) {
     const Component = this._currentElement.type;
     let instanceOrElement;
     if (doConstruct) {
