@@ -17,7 +17,7 @@ class CylinderGeometryDescriptor extends GeometryDescriptorBase {
       'openEnded',
       'thetaStart',
       'thetaLength',
-    ].forEach(propName => {
+    ].forEach((propName) => {
       this.hasProp(propName, {
         type: PropTypes.number,
         update: this.triggerRemount,
