@@ -15,7 +15,7 @@ class RingGeometryDescriptor extends GeometryDescriptorBase {
       'phiSegments',
       'thetaStart',
       'thetaLength',
-    ].forEach(propName => {
+    ].forEach((propName) => {
       this.hasProp(propName, {
         type: PropTypes.number,
         update: this.triggerRemount,
