@@ -13,7 +13,7 @@ class AbsArcDescriptor extends ShapeActionDescriptorBase {
       'radius',
       'startAngle',
       'endAngle',
-    ].forEach(propName => {
+    ].forEach((propName) => {
       this.hasProp(propName, {
         type: PropTypes.number.isRequired,
         update: this.triggerRemount,
