@@ -12,7 +12,7 @@ class QuadraticCurveToDescriptor extends ShapeActionDescriptorBase {
       'cpY',
       'x',
       'y',
-    ].forEach(propName => {
+    ].forEach((propName) => {
       this.hasProp(propName, {
         type: PropTypes.number.isRequired,
         update: this.triggerRemount,
