@@ -215,7 +215,7 @@ class THREEElementDescriptor {
     threeObject.userData.events = eventsForObject;
     threeObject.userData._descriptor = this;
 
-    this._updateInitial.forEach(propertyName => {
+    this._updateInitial.forEach((propertyName) => {
       if (props.hasOwnProperty(propertyName)) {
         this.propUpdates[propertyName](threeObject, props[propertyName], true);
       } else {
@@ -229,7 +229,7 @@ class THREEElementDescriptor {
       }
     });
 
-    this._simpleProperties.forEach(propertyName => {
+    this._simpleProperties.forEach((propertyName) => {
       if (props.hasOwnProperty(propertyName)) {
         threeObject[propertyName] = props[propertyName];
       }
@@ -338,7 +338,7 @@ class THREEElementDescriptor {
   }
 
   registerSimpleProperties(propertyNames) {
-    propertyNames.forEach(propName => {
+    propertyNames.forEach((propName) => {
       if (this._simpleProperties.indexOf(propName) === -1) {
         this._simpleProperties.push(propName);
       }
